Move triangle to new pkg
