+++ pkg/primitives/triangle/README.md
@@ -1,0 +1,1 @@
+# Triangle
