+++ pkg/primitives/triangle/triangle.go
@@ -1,4 +1,4 @@
-package primitives
+package triangle
 
 import (
 	"github.com/go-gl/gl/v4.1-core/gl"
