+++ pkg/primitives/triangle/triangle_test.go
@@ -1,0 +1,5 @@
+package triangle
+
+import (
+	"testing"
+)
